Module rslogic::goal

Logical Goals

Goals are used to specify logical statements.

Structs

Conjunction
A Conjunction goal evaluates its sub-goal a using a given state, then evaluates sub-goal b using the results.

Disjunction
Evaluating a Disjunction goal returns all the possible states of evaluating a and b.

Fail
Evaluating a Fail goal always results in zero states.

Predicate
Evaluating a Predicate goal returns the given state only if the function returns true.

UnifyVal
Evaluating a UnifyVal goal attempts to unify a variable and a value.

UnifyVar
Evaluating a UnifyVar goal attempts to unify the variables.

Traits

Goal
Evaluate a Goal to produce zero or more States, or collections of variable bindings.

Functions

conj
Creates a Conjunction goal which returns the conjunction (logical AND) of evaluating the two sub-goals.

disj
Creates a Disjunction goal which returns the disjunction (logical OR) of evaluating the two sub-goals.

fail
Creates a Fail goal.

pred
Creates a Predicate goal that filters a set of possible states with the given function.

unify_val
Creates a UnifyVal goal that attempts to unify the variable and the value.

unify_vars
Creates a UnifyVar goal that attempts to unify the variables.