Trait rslogic::state::Unif

pub trait Unif<T> where T: PartialEq + Unif<T> {
    fn unify(&self, other: &T, prev: &State<T>) -> PossibleStates<T>;
}

Values used in a state must be unifiable. Unifying two values produces zero or more possible states, where variables that may be contained in the values may be bound in various combinations.

Required Methods

fn unify(&self, other: &T, prev: &State<T>) -> PossibleStates<T>